+++ src/main/java/com/OnlineStore/OnlineStore/models/entity/OrderItem.java
@@ -49,7 +49,7 @@ public class OrderItem implements Serializable {
 		this.quantity = quantity;
 	}
 
-	public Double calculateRow() {
+	public float calculateRow() {
 		return product.getPrice()*quantity;
 	}
 }
